+++ projects/swimlane/ngx-charts/src/lib/models/chart-data.model.ts
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/no-empty-interface */
-
 import { IVector2D } from './coordinates.model';
 
 export type StringOrNumberOrDate = string | number | Date;